+++ client/src/pages/dashboard/components/index.js
@@ -1,2 +1,1 @@
-export { default as Create } from "./Moto_Create/Create";
 export { default as Moto_Create } from "./Moto_Create/Moto_Create";
