+++ client/src/pages/dashboard/index.js
@@ -1,5 +1,4 @@
 export { default as AppDashboard } from "./AppDashboard";
-export { default as Create } from "./components/Moto_Create/Create";
 export { default as Moto_Create } from "./components/Moto_Create/Moto_Create";
 export * from "./components";
 export * from "./helper";
