+++ client/src/pages/dashboard/pages/Dashboard.jsx
@@ -1,11 +1,11 @@
 import React from "react";
-import { Create } from "../components";
+import { Moto_Create } from "../components";
 
 const Dashboard = () => {
   return (
     <div className=" flex m-auto my-10 min-w-[100%] justify-center">
       <div className="w-full">
-        <Create />
+        <Moto_Create />
       </div>
     </div>
   );
